LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-name-shadow-expected.txt


FAIL Outer animation can see view timeline defined by :host assert_equals: expected 1 but got 0
FAIL Outer animation can see view timeline defined by ::slotted assert_equals: expected 1 but got 0
FAIL Inner animation can see view timeline defined by ::part assert_equals: expected 1 but got 0
FAIL Slotted element can see view timeline within the shadow assert_equals: expected 1 but got 0
